ReiDBG: Rename from Reibug, bug fixes
