+++ reidbg/Properties/AssemblyInfo.cs
